requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
gtts
supabase
openai
google-genai